Update facade.py and UserService.py & Create ressources.py
Move function is_valid_uuid4 to a separate ressources module
