+++ part2/hbnb/app/services/facade.py
@@ -4,16 +4,6 @@ from app.models.place import Place
 from app.services.UserService import UserService
 from app.models.review import Review
 from app.models.baseEntity import type_validation
-from uuid import UUID
-
-
-def is_valid_uuid4(uuid_str):
-    """Determines if given str is a uuid4"""
-    try:
-        val = UUID(uuid_str, version=4)
-        return val.version == 4
-    except ValueError:
-        return False
 
 
 class HBnBFacade:
